fix: Nueva version 6.8.2 Entrego la nueva versión liberada por el ministerio en Enero de 2026

+++ fuentes/eeutil-vis-docexp/src/main/java/es/mpt/dsic/inside/fop/converter/TemplateConverter.java
@@ -1,12 +1,12 @@
 /*
- * Copyright (C) 2012-13 MINHAP, Gobierno de España This program is licensed and may be used,
- * modified and redistributed under the terms of the European Public License (EUPL), either version
- * 1.1 or (at your option) any later version as soon as they are approved by the European
- * Commission. Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language governing permissions and
- * more details. You should have received a copy of the EUPL1.1 license along with this program; if
- * not, you may find it at http://joinup.ec.europa.eu/software/page/eupl/licence-eupl
+ * Copyright (C) 2025, Gobierno de España This program is licensed and may be used, modified and
+ * redistributed under the terms of the European Public License (EUPL), either version 1.1 or (at
+ * your option) any later version as soon as they are approved by the European Commission. Unless
+ * required by applicable law or agreed to in writing, software distributed under the License is
+ * distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied. See the License for the specific language governing permissions and more details. You
+ * should have received a copy of the EUPL1.1 license along with this program; if not, you may find
+ * it at http://joinup.ec.europa.eu/software/page/eupl/licence-eupl
  */
 
 package es.mpt.dsic.inside.fop.converter;
@@ -18,14 +18,10 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
+
 import javax.xml.bind.JAXBException;
 import javax.xml.parsers.ParserConfigurationException;
-import net.sf.jasperreports.engine.JREmptyDataSource;
-import net.sf.jasperreports.engine.JasperCompileManager;
-import net.sf.jasperreports.engine.JasperExportManager;
-import net.sf.jasperreports.engine.JasperFillManager;
-import net.sf.jasperreports.engine.JasperPrint;
-import net.sf.jasperreports.engine.JasperReport;
+
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -34,12 +30,21 @@ import org.apache.pdfbox.multipdf.PDFMergerUtility;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.xml.sax.SAXException;
+
+import es.mpt.dsic.inside.configure.ConfigureRestInfo;
 import es.mpt.dsic.inside.fop.model.documento.DocumentoEniConMAdicionales;
 import es.mpt.dsic.inside.pdf.converter.PdfConverter;
-import es.mpt.dsic.inside.pdf.exception.PdfConversionException;
 import es.mpt.dsic.inside.security.context.AplicacionContext;
+import es.mpt.dsic.inside.utils.exception.EeutilException;
 import es.mpt.dsic.inside.utils.file.FileUtil;
 import es.mpt.dsic.inside.ws.service.model.DocumentoContenido;
+import net.sf.jasperreports.engine.JREmptyDataSource;
+import net.sf.jasperreports.engine.JRException;
+import net.sf.jasperreports.engine.JasperCompileManager;
+import net.sf.jasperreports.engine.JasperExportManager;
+import net.sf.jasperreports.engine.JasperFillManager;
+import net.sf.jasperreports.engine.JasperPrint;
+import net.sf.jasperreports.engine.JasperReport;
 
 @Component
 public class TemplateConverter {
@@ -49,7 +54,7 @@ public class TemplateConverter {
 
   private static final String CONTENT_DATA_TEMP = "Content_data";
 
-  protected final static Log logger = LogFactory.getLog(TemplateConverter.class);
+  protected static final Log logger = LogFactory.getLog(TemplateConverter.class);
 
   @Autowired
   PdfConverter pdfConverter;
@@ -66,37 +71,53 @@ public class TemplateConverter {
    * @throws ParserConfigurationException
    */
   public DocumentoContenido convertDocumentToPdf(DocumentoEniConMAdicionales docEniAdicionales,
-      byte[] jrxml) throws Exception {
+      byte[] jrxml, String idApp) throws EeutilException {
     DocumentoContenido retorno = new DocumentoContenido();
 
-    JasperReport report = JasperCompileManager.compileReport(new ByteArrayInputStream(jrxml));
+    try (ByteArrayInputStream bArrayIn = new ByteArrayInputStream(jrxml)) {
 
-    Map<String, Object> params = new HashMap<String, Object>();
-    params.put("data", docEniAdicionales);
+      JasperReport report = JasperCompileManager.compileReport(bArrayIn);
 
-    JasperPrint print = JasperFillManager.fillReport(report, params, new JREmptyDataSource());
+      Map<String, Object> params = new HashMap<>();
+      params.put("data", docEniAdicionales);
 
-    ByteArrayOutputStream bout = new ByteArrayOutputStream();
-    JasperExportManager.exportReportToPdfStream(print, bout);
+      JasperPrint print = JasperFillManager.fillReport(report, params, new JREmptyDataSource());
 
-    // conversion del contenido a pdf
-    byte[] contenidoConvertido =
-        convertContent(docEniAdicionales.getDocumento().getContenido().getValorBinario());
+      try (ByteArrayOutputStream bout = new ByteArrayOutputStream();
+          ByteArrayOutputStream pdfSalida = new ByteArrayOutputStream();) {
+        JasperExportManager.exportReportToPdfStream(print, bout);
 
-    ByteArrayOutputStream pdfSalida = new ByteArrayOutputStream();
-    PDFMergerUtility merger = new PDFMergerUtility();
-    merger.setDestinationStream(pdfSalida);
-    merger.addSource(new ByteArrayInputStream(bout.toByteArray()));
-    merger.addSource(new ByteArrayInputStream(contenidoConvertido));
-    merger.mergeDocuments(MemoryUsageSetting.setupTempFileOnly());
+        // conversion del contenido a pdf
+        byte[] contenidoConvertido = convertContent(
+            docEniAdicionales.getDocumento().getContenido().getValorBinario(), idApp);
 
-    retorno.setBytesDocumento(pdfSalida.toByteArray());
-    retorno.setMimeDocumento("application/pdf");
+        PDFMergerUtility merger = new PDFMergerUtility();
+        merger.setDestinationStream(pdfSalida);
+
+        try (ByteArrayInputStream bArrayIn2 = new ByteArrayInputStream(bout.toByteArray());
+            ByteArrayInputStream bArrayIn3 = new ByteArrayInputStream(contenidoConvertido)) {
+          merger.addSource(bArrayIn2);
+          merger.addSource(bArrayIn3);
+          merger.mergeDocuments(MemoryUsageSetting.setupTempFileOnly());
+        }
+
+        // retorno.setBytesDocumento(Arrays.copyOf(pdfSalida.toByteArray(),pdfSalida.toByteArray().length));
+        retorno.setBytesDocumento(pdfSalida.toByteArray());
+      }
+      retorno.setMimeDocumento("application/pdf");
+
+    } catch (IOException | JRException e) {
+      throw new EeutilException(e.getMessage(), e);
+    } catch (EeutilException e) {
+      throw new EeutilException(e.getMessage(), e);
+    } catch (Exception e) {
+      throw new EeutilException(e.getMessage(), e);
+    }
 
     return retorno;
   }
 
-  private byte[] convertContent(byte[] contenido) {
+  private byte[] convertContent(byte[] contenido, String idApp) throws EeutilException {
     byte[] retorno = null;
     File converted = null;
     try {
@@ -105,22 +126,33 @@ public class TemplateConverter {
       String ipOO = aplicacionContext.getAplicacionInfo().getPropiedades().get("ip.openoffice");
       String portOO = aplicacionContext.getAplicacionInfo().getPropiedades().get("port.openoffice");
 
-      converted = pdfConverter.convertir(ipOO, portOO, fileIn, null);
+      converted = pdfConverter.convertir(ipOO, portOO, fileIn, null, idApp,
+          ConfigureRestInfo.getBaseUrlRestIgae(), ConfigureRestInfo.getTokenNameRestIgae(),
+          ConfigureRestInfo.getTokenValueRestIgae());
 
       retorno = FileUtils.readFileToByteArray(converted);
-    } catch (NumberFormatException e) {
-      e.printStackTrace();
-    } catch (IOException e) {
-      e.printStackTrace();
-    } catch (PdfConversionException e) {
-      e.printStackTrace();
-    } finally {
+    } catch (NumberFormatException | IOException | EeutilException e) {
+      // logger.error("Fallo al convertir",e);
+      throw new EeutilException("Fallo al convertir" + e.getMessage(), e);
+    } catch (Exception e) {
+      // logger.error("Fallo al convertir",e);
+      throw new EeutilException("Fallo al convertir" + e.getMessage(), e);
+    }
+
+    finally {
       try {
-        FileUtils.forceDelete(converted);
+        if (converted != null && converted.exists()) {
+          FileUtils.forceDelete(converted);
+        }
       } catch (IOException e) {
-        logger.error("Error al eliminar fichero temporal:" + converted.getAbsolutePath());
+        // logger.error("Error al eliminar fichero temporal:" + converted.getAbsolutePath());
+        throw new EeutilException("Error al eliminar fichero temporal:"
+            + converted.getAbsolutePath() + " " + e.getMessage(), e);
       }
     }
+
+
+
     return retorno;
   }
 
@@ -131,20 +163,15 @@ public class TemplateConverter {
    * @return
    * @throws IOException
    */
-  private File createFile(byte[] content) throws IOException {
-    FileOutputStream fos = null;
-    try {
-      String filePathIn = FileUtil.createFilePath(CONTENT_DATA_TEMP);
-      fos = new FileOutputStream(filePathIn);
+  private File createFile(byte[] content) throws EeutilException {
+
+
+    String filePathIn = FileUtil.createFilePath(CONTENT_DATA_TEMP);
+    try (FileOutputStream fos = new FileOutputStream(filePathIn);) {
       fos.write(content);
-      fos.close();
       return new File(filePathIn);
     } catch (IOException e) {
-      throw e;
-    } finally {
-      if (fos != null) {
-        fos.close();
-      }
+      throw new EeutilException(e.getMessage(), e);
     }
   }
